move images out of viewmodel
