+++ AvaGui/ViewModels/SubObjectTypes/ImageTableViewModel.cs
@@ -1,6 +1,4 @@
-using Avalonia;
 using Avalonia.Media.Imaging;
-using Avalonia.Platform;
 using OpenLoco.Dat;
 using OpenLoco.Dat.Types;
 using ReactiveUI;
@@ -47,10 +45,11 @@ namespace AvaGui.ViewModels
 	{
 		readonly ILocoObject Parent;
 
-		public ImageTableViewModel(ILocoObject parent, PaletteMap paletteMap)
+		public ImageTableViewModel(ILocoObject parent, PaletteMap paletteMap, IList<Bitmap> images)
 		{
 			Parent = parent;
 			PaletteMap = paletteMap;
+			Images = images;
 
 			_ = this.WhenAnyValue(o => o.Parent)
 				.Subscribe(_ => this.RaisePropertyChanged(nameof(Images)));
@@ -130,18 +129,8 @@ namespace AvaGui.ViewModels
 		[Reactive]
 		public int Zoom { get; set; } = 1;
 
-		public List<Bitmap> Images
-		{
-			get
-			{
-				images = CreateImages(Parent.G1Elements, PaletteMap, Zoom).ToList();
-				return images;
-			}
-			set =>
-				//images = value;
-				_ = this.RaiseAndSetIfChanged(ref images, value);
-		}
-		List<Bitmap> images;
+		[Reactive]
+		public IList<Bitmap> Images { get; set; }
 
 		[Reactive]
 		public int SelectedImageIndex { get; set; } = -1;
@@ -173,156 +162,8 @@ namespace AvaGui.ViewModels
 			return "<unk>";
 		}
 
-		public static IEnumerable<Bitmap> CreateImages(IEnumerable<G1Element32> g1Elements, PaletteMap paletteMap, int zoom)
-		{
-			foreach (var g1Element in g1Elements)
-			{
-				if (g1Element.ImageData.Length == 0)
-				{
-					//logger?.Info($"skipped loading g1 element {i} with 0 length");
-					continue;
-				}
 
-				if (g1Element.Flags.HasFlag(G1ElementFlags.IsR8G8B8Palette))
-				{
-					yield return G1RGBToBitmap(g1Element, zoom);
-				}
-				else
-				{
-					yield return G1IndexedToBitmap(g1Element, paletteMap, true, zoom);
-				}
-			}
-		}
 
-		static Bitmap G1RGBToBitmap(G1Element32 g1Element, int zoom = 1)
-		{
-			var imageData = g1Element.ImageData;
-			var writeableBitmap = new WriteableBitmap(
-				new PixelSize(g1Element.Width * zoom, g1Element.Height * zoom),
-				new Vector(96, 96),  // DPI
-				PixelFormat.Rgba8888); // Or a suitable pixel format
 
-			using (var lockedBitmap = writeableBitmap.Lock())
-			{
-				unsafe
-				{
-					var pointer = (uint*)lockedBitmap.Address; // Access pixel data directly
-
-					for (var y = 0; y < g1Element.Height; y++)
-					{
-						for (var x = 0; x < g1Element.Width; x++)
-						{
-							// Calculate pixel index
-							var index = x + (y * g1Element.Width);
-
-							// Set pixel color (example: red)
-							pointer[index] = 0xFFFF0000;
-						}
-					}
-				}
-			}
-
-			return writeableBitmap;
-		}
-
-		static Bitmap G1IndexedToBitmap(G1Element32 g1Element, PaletteMap paletteMap, bool useTransparency = false, int zoom = 1)
-		{
-			var writeableBitmap = new WriteableBitmap(
-				new PixelSize(g1Element.Width, g1Element.Height),
-				new Vector(96, 96),  // DPI
-				PixelFormat.Rgba8888); // Or a suitable pixel format
-
-			using (var lockedBitmap = writeableBitmap.Lock())
-			{
-				unsafe
-				{
-					var ptr = (byte*)lockedBitmap.Address; // Access pixel data directly
-
-					for (var y = 0; y < g1Element.Height; y++)
-					{
-						for (var x = 0; x < g1Element.Width; x++)
-						{
-							var index = x + (y * g1Element.Width);
-							var paletteIndex = g1Element.ImageData[index];
-
-							if (paletteIndex == 0 && useTransparency)
-							{
-								ptr += 4;
-							}
-							else
-							{
-								var colour = paletteMap.Palette[paletteIndex].Color;
-								var pixel = colour.ToPixel<Rgb24>();
-
-								//var ptr = (byte*)pointer;
-								*ptr++ = pixel.R;
-								*ptr++ = pixel.G;
-								*ptr++ = pixel.B;
-								*ptr++ = 255;
-							}
-						}
-					}
-				}
-			}
-
-			// bug in avalonia/skiasharp/skia: https://github.com/AvaloniaUI/Avalonia/issues/8444
-			//return writeableBitmap.CreateScaledBitmap(new PixelSize(g1Element.Width * zoom, g1Element.Height * zoom), BitmapInterpolationMode.None);
-
-			return writeableBitmap;
-		}
-
-		//static Bitmap G1IndexedToBitmapScaled(G1Element32 g1Element, PaletteMap paletteMap, bool useTransparency = false, int zoom = 1)
-		//{
-		//	var info = new SKImageInfo(g1Element.Width, g1Element.Height, SKColorType.Rgba8888, SKAlphaType.Opaque);
-		//	var img = SKImage.Create(info);
-		//	var bmp = SKBitmap.FromImage(img);
-
-		//	unsafe
-		//	{
-		//		var ptr = (uint*)bmp.GetPixels();
-		//		for (var y = 0; y < g1Element.Height; y++)
-		//		{
-		//			for (var x = 0; x < g1Element.Width; x++)
-		//			{
-		//				var index = x + (y * g1Element.Width);
-		//				var paletteIndex = g1Element.ImageData[index];
-
-		//				if (paletteIndex == 0 && useTransparency)
-		//				{
-		//					ptr += 4;
-		//				}
-		//				else
-		//				{
-		//					var colour = paletteMap.Palette[paletteIndex].Color;
-		//					var pixel = colour.ToPixel<Rgb24>();
-
-		//					//var ptr = (byte*)pointer;
-		//					*ptr++ = pixel.R;
-		//					*ptr++ = pixel.G;
-		//					*ptr++ = pixel.B;
-		//					*ptr++ = 255;
-		//				}
-		//			}
-		//		}
-		//	}
-
-		//	var scaledImage = new SKBitmap(g1Element.Width * zoom, g1Element.Height * zoom, SKColorType.Rgba8888, SKAlphaType.Opaque);
-		//	_ = bmp.ScalePixels(scaledImage, SKFilterQuality.None);
-
-		//	// Encode the SKBitmap into a memory stream (using PNG format for best compatibility)
-		//	using (var memoryStream = new MemoryStream())
-		//	{
-		//		_ = scaledImage.Encode(memoryStream, SKEncodedImageFormat.Png, 100); // 100 is the quality (0-100)
-
-		//		// Create an Avalonia Bitmap from the memory stream
-		//		memoryStream.Position = 0;
-		//		var avaloniaBitmap = new Bitmap(memoryStream);
-		//		return avaloniaBitmap;
-		//	}
-		//	// bug in avalonia/skiasharp/skia: https://github.com/AvaloniaUI/Avalonia/issues/8444
-		//	//return writeableBitmap.CreateScaledBitmap(new PixelSize(g1Element.Width * zoom, g1Element.Height * zoom), BitmapInterpolationMode.None);
-
-		//	//return writeableBitmap;
-		//}
 	}
 }
